+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## next
 
+- Added `spaceBytes` field to `stringifyInfo()` result, which indicates the number of bytes used for white spaces. This allows for estimating size of `JSON.stringify()` result with and without formatting (when `space` option is used) in a single pass instead of two
 - Fixed `stringifyInfo()` to correctly accept the `space` parameter from options, i.e. `stringifyInfo(data, { space: 2 })`
 
 ## 0.6.1 (2024-08-06)